% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm_movement.R
\docType{data}
\name{arm_movement}
\alias{arm_movement}
\title{Arm Movement and Mood}
\format{
A data frame with 56 observations of 11 variables.
\describe{
  \item{id}{nonconsecutive integers from 1-67}
  \item{datetime}{time of the experiment as datetime}
  \item{baseline_mean}{mean score of characters under neutral conditions}
  \item{passive_*}{difference in score of characters when arm was passively moving in direction indicated}
  \item{active_*}{difference in score of characters when arm was actively moving in direction indicated}
  \item{sex}{0=male, 1=female}
  \item{age}{integer, years}
}

From the authors: "Previous research showed that the activation of the arm extensor muscles is associated with a more negative attitude. In contrast, arm flexor muscle activation is associated with a more positive one. However, most of the studies tested these effects with isometric activation of the muscles. The current study was conducted to test if these effects occur with actual passive and active forearm movement."
}
\source{
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0327881
}
\usage{
arm_movement
}
\description{
Data from an experiment testing measuring preferences toward Chinese-like ideographs while their arm is moving in various directions, either actively or passively.
}
\keyword{datasets}
